+++ CHANGELOG.md
@@ -7,4 +7,6 @@ Added Bagrut grades and Hatamot APIs.
 ## 0.0.4
 Added Api enum and Data Processor to Api Controller.
 ## 0.0.5
-put all models file in one file, and by doing this exported another two models that were not exported in version 0.0.4
+Put all models file in one file, and by doing this exported another two models that were not exported in version 0.0.4
+## 0.0.6
+Added statusCode to Login, and some nice getters such as isUnauthorized.